derp derp module 8

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,28 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
-            _context = context;
+	        _context = context;
+	        _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+	        var signedInUser = _userManager.GetUserAsync(HttpContext.User);
+            var model = _context.Items.Where(x => x.User.Id == signedInUser.Result.Id).ToList();
 
             return View("Index", model);
         }
@@ -27,9 +34,12 @@ namespace WishList.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(Models.Item item)
+        public IActionResult Create(Item item)
         {
-            _context.Items.Add(item);
+	        var signedInUser = _userManager.GetUserAsync(HttpContext.User).Result;
+
+	        _context.Items.Add(signedInUser.Items.FirstOrDefault());
+            //_context.Items.Add(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
         }
@@ -37,7 +47,17 @@ namespace WishList.Controllers
         public IActionResult Delete(int id)
         {
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
-            _context.Items.Remove(item);
+            var signedInUser = _userManager.GetUserAsync(HttpContext.User).Result;
+
+            if (item.User == signedInUser)
+            {
+                _context.Items.Remove(item);
+            }
+            else
+            {
+	            return Unauthorized();
+            }
+            
             _context.SaveChanges();
             return RedirectToAction("Index");
         }
